+++ src/components/FlawForm/FlawForm.vue
@@ -479,7 +479,7 @@ const toggleComponentsTooltip = () => {
             @updateFlaw="updateFlaw"
             @update:model-value="onUnembargoed"
           />
-          <FlawFormOwner v-model="flaw.owner" :taskKey="flaw.task_key" />
+          <FlawFormOwner v-model="flaw.owner" :taskKey="flaw.task_key" :error="errors.owner" />
           <LabelStatic
             v-if="mode === 'edit'"
             :modelValue="createdDate"